Ergotism symptoms

A dataset containing the ergotism data presented in Spanos and Gottlieb (1976) Table 1. The dataframe records the symptoms possible symptoms of ergotism reported about 22 individuals.

ergotism_symptoms

Format

A data frame with 22 rows and 20 variables:

name

Name of the sufferer

rsw_volume

Volume number of RSW

rsw_page

Page number

vomitting

Vomitting

diarrhea

Diarrhea

livid_skin

Livid skin

permanent_contractures

Permanent contractures

pain_in_extremities

Pain in extremities

death

Death

temporary_muscle_stiffness

Temporary muscle stiffness

convulsions

Convulsions Cases Spanos and Gottlieb considered uncertain are coded missing.

ravenous_appetite

Ravenous appetite

perceptual_disturbances

Perceptual disturbances Cases Spanos and Gottlieb considered uncertain are coded missing.

apparitions

Apparitions

sensations_of_hot_and_cold

Sensations of hot & cold

skin_sensations

Skin sensations

stomach_pain

Stomach pain

choking_sensations

Choking sensations

inability_to_speak

Temporary inability to speak

total

Total number of symptoms

Source

Spanos, N. P., & Gottlieb, J. (1976). Ergotism and the Salem Village Witch Trials. Science, 194(4272), 1390–1394.

Details

Spanos and Gottlieb, with one exeption (E. Keysar), drew the data from depositions in the Records of the Salem Witchcraft collection. Details of data collection are provided in the article. The data set indicates the volume and page of the individual report. J. Doritch is also mentioned in volume 2 page 179 of RSW.

The article is an attempt to compare the actual symptoms reported to a list of symptoms of ergotism. For each variable a 1 inidicates that the symptom was reported, while a 0 indicated it was not reported. They conclude that "We examined these 112 depositions looking for behavior that, even in a broad sense, might possibly represent symptoms of convulsive ergotism."