Simple Features data for Salem

A dataset containing map information for the three counties represented in the accused_witches data set (not including present-day Maine locations) in simple features format. It also includes associated data on accusations. This data should never be used to analyze accusations. The data is based on modern town boundaries from the Massachuessetts Document Repository. Some simple cleaning has been done to facilitate basic mapping. Specifically, the names Salem Town and Salem Village are used rather than the modern Salem and Danvers. Additionally, a field, TOWN_LABEL is attached to the largest area for each town or community.

salem_region

Format

A data frame with 437 rows and 28 variables:

TOWN

Name of town in upper case letters

TOWN_ID

ID of town

TYPE

Type of municipality C = City or T = Town

ISLAND

Indicator for islands. Note that many towns have multiple polygons

COASTAL_PO

YES for ocean polygon within town limits, NO otherwise

FOURCOLOR

Codes allowing shading of all towns with 4 symbols

FIPS_STCO

FIPS for state and county

CCD_MCD

US Census town code

FIPS_PLACE

FIPS town code

FIPS_MCD

FIPS State & County & Census Town Code concatenated

FIPS_COUNT

FIPS County

ACRES

Area in acres

SQUARE_MIL

Area in square miles

SHAPE_Leng

The perimeter in meters

SHAPE_Area

The polygon area in square meters

n_accused

Total accused

February

Number accused in February

March

Number accused in March

April

Number accused in April

May

Number accused in May

June

Number accused in June

July

Number accused in July

August

Number accused in August

September

Number accused in September

October

Number accused in October

November

Number accused in November

TOWN_LABEL

Town name to be used for labelling

geometry

Simple Features geometry

Source

https://docs.digital.mass.gov/dataset/massgis-data-community-boundaries-towns-survey-points

Details

Complex analyses require the sf package, but simple mapping can be done using ggplot2.

FIPS refers to the Federal Information Processing Standard